Remove default implementations for filter apis
