+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/v17/BitcoindV17RpcClient.scala
@@ -8,6 +8,7 @@ import org.bitcoins.commons.jsonmodels.bitcoind.{
 }
 import org.bitcoins.commons.serializers.JsonSerializers._
 import org.bitcoins.commons.serializers.JsonWriters._
+import org.bitcoins.core.api.ChainQueryApi
 import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.core.script.crypto.HashType
 import org.bitcoins.crypto.ECPrivateKey
@@ -37,6 +38,18 @@ class BitcoindV17RpcClient(override val instance: BitcoindInstance)(
     with V17PsbtRpc {
 
   override def version: BitcoindVersion = BitcoindVersion.V17
+
+  override def getFilterCount: Future[Int] =
+    Future.failed(
+      throw new UnsupportedOperationException(
+        s"bitcoind ${instance.getVersion} does not support block filters"))
+
+  override def getFiltersBetweenHeights(
+      startHeight: Int,
+      endHeight: Int): Future[Vector[ChainQueryApi.FilterResponse]] =
+    Future.failed(
+      throw new UnsupportedOperationException(
+        s"bitcoind ${instance.getVersion} does not support block filters"))
 
   /**
     * $signRawTx
