+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/v18/BitcoindV18RpcClient.scala
@@ -7,6 +7,7 @@ import org.bitcoins.commons.jsonmodels.bitcoind.{
 }
 import org.bitcoins.commons.serializers.JsonSerializers._
 import org.bitcoins.commons.serializers.JsonWriters._
+import org.bitcoins.core.api.ChainQueryApi
 import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.core.script.crypto.HashType
 import org.bitcoins.crypto.ECPrivateKey
@@ -36,6 +37,18 @@ class BitcoindV18RpcClient(override val instance: BitcoindInstance)(
     with V18AssortedRpc {
 
   override lazy val version: BitcoindVersion = BitcoindVersion.V18
+
+  override def getFilterCount: Future[Int] =
+    Future.failed(
+      throw new UnsupportedOperationException(
+        s"bitcoind ${instance.getVersion} does not support block filters"))
+
+  override def getFiltersBetweenHeights(
+      startHeight: Int,
+      endHeight: Int): Future[Vector[ChainQueryApi.FilterResponse]] =
+    Future.failed(
+      throw new UnsupportedOperationException(
+        s"bitcoind ${instance.getVersion} does not support block filters"))
 
   /**
     * $signRawTx
